+++ src/common/display.ts
@@ -73,7 +73,6 @@ interface Line {
  * @todo fix RTL unicode selection display
  * @todo Add MXP image height - requires variable line height support
  * @todo Add/fox MXP image selection highlighting
- * @todo Add enable colors/background colors - let them be turned on and off quickly, use the new styles rule to add .view > span span {color: inherit !important;} and simalrot background
  */
 export class Display extends EventEmitter {
     private _lineID = 0;
@@ -155,6 +154,22 @@ export class Display extends EventEmitter {
 
     private _showSplitButton = true;
     private _hideTrailingEmptyLine = true;
+    private _enableColors = true;
+    private _enableBackgroundColors = true;
+
+    get enableColors() { return this._enableColors; }
+    set enableColors(value) {
+        if (value === this._enableColors) return;
+        this._enableColors = value;
+        this.buildStyleSheet();
+    }
+
+    get enableBackgroundColors() { return this._enableBackgroundColors; }
+    set enableBackgroundColors(value) {
+        if (value === this._enableBackgroundColors) return;
+        this._enableBackgroundColors = value;
+        this.buildStyleSheet();
+    }
 
     get showSplitButton() { return this._showSplitButton; }
     set showSplitButton(value) {
@@ -397,6 +412,10 @@ export class Display extends EventEmitter {
     constructor(display: string | JQuery | HTMLElement, options?);
     constructor(display?: any, options?: DisplayOptions) {
         super();
+        if (options && options.hasOwnProperty('backgroundColors'))
+            this._enableBackgroundColors = options.backgroundColors;
+        if (options && options.hasOwnProperty('colors'))
+            this._enableColors = options.colors;
         if (!display)
             throw new Error('Display must be an id, element or jquery object');
         if (typeof display === 'string') {
@@ -465,7 +484,7 @@ export class Display extends EventEmitter {
 
         this._charHeight = $(this._character).innerHeight();
         this._charWidth = parseFloat(window.getComputedStyle(this._character).width);
-        this._styles.innerHTML = `.background > span, .view > span, .line, .background-line { height: ${this._charHeight}px; }`;
+        this.buildStyleSheet();
 
         this._VScroll = new ScrollBar(this._el, this._view);
         this._VScroll.on('scroll', () => {
@@ -1141,7 +1160,7 @@ export class Display extends EventEmitter {
             //recalculate height/width of characters so display can be calculated
             this._charHeight = $(this._character).innerHeight();
             this._charWidth = parseFloat(window.getComputedStyle(this._character).width);
-            this._styles.innerHTML = `.background > span, .view > span, .line, .background-line { height: ${this._charHeight}px; }`;
+            this.buildStyleSheet();
             /*
             let html = this._htmlLines, t;
             let h = this._charHeight;
@@ -2266,6 +2285,15 @@ export class Display extends EventEmitter {
             this._padding = padding;
             this.doUpdate(UpdateType.view | UpdateType.selection | UpdateType.scrollbars);
         }
+    }
+
+    private buildStyleSheet() {
+        let styles = `.background > span, .view > span, .line, .background-line { height: ${this._charHeight}px; }`;
+        if (!this._enableColors)
+            styles += '.view > span span {color: inherit !important;}';
+        if (!this._enableColors || !this._enableBackgroundColors)
+            styles += '.background > span span {background-color: inherit !important;}';
+        this._styles.innerHTML = styles;
     }
 
     private buildLineDisplay(idx?: number) {