+++ changelog.md
@@ -4,6 +4,8 @@
 
 - **New:**
   - Added [Hide trailing line](docs/preferences.md#display) display preference to hide trailing empty line.
+  - Added [Enable colors](docs/preferences.md#display) enable or disable colors.
+  - Added [Enable background colors](docs/preferences.md#display) enable or disable background colors
   - Added new Zen theme
   - ANSI Parser: Added backspace character support
 - **Changed:**
@@ -14,6 +16,7 @@
   - Logger:
     - Re-coded html logger to use classes and style blocks to reduce file sizes
     - Convert write system to use a stream to batch write to improve write speeds
+  - Preferences: Move display font selection to own sub section under display
 - **Fixed:**
   - Display:
     - Detect changes to display better to handle theme changes